added unit count check, removed comments at start, pushing to be able to pull latest
